requirements.txt
langchain>=0.1.0
langchain-google-genai>=1.0.0
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1
langchain-pinecone>=0.1.0
google-generativeai>=0.3.0
chromadb>=0.4.0
pinecone-client>=3.0.0
python-dotenv>=1.0.0
